Smarter determination of loglevel.
If not specified, use the current level of the root logger.  This assumes
that you've configured logging before you wrap your app in the Logview
middleware.  I'm not sure if there's a better way...

Note, that the current paster entry points/our setup() don't let you
actually specify a different log level.  That's probably a bug...